+++ app/src/main/AndroidManifest.xml
@@ -1,29 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="demourak0.rowan.edu.badger">
+    package="seniorproject.badger">
 
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
-        android:supportsRtl="true"
-        android:theme="@style/AppTheme">
-        <activity android:name=".ViewFriendsActivity">
+        android:supportsRtl="true">
+        <activity android:name=".Login"
+                  android:theme="@style/Theme.AppCompat.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".LoginActivity" />
-        <activity android:name=".ViewProfileActivity" />
-        <activity
-            android:name=".SettingsActivity"
-            android:label="@string/title_activity_settings" />
-        <activity android:name=".CreateBadgeActivity" />
-        <activity android:name=".BadgeDisplayActivity" />
-        <activity android:name=".CreateGroupActivity" />
-        <activity android:name=".ViewGroupsActivity"></activity>
     </application>
+
 
 </manifest>
