--- app/src/main/java/demourak0/rowan/edu/badger/ViewFriendsActivity.java
@@ -1,13 +1,0 @@
-package demourak0.rowan.edu.badger;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class ViewFriendsActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_view_friends);
-    }
-}
